Extra warnings added and fixes to make them clear (#223)
* Extra warnings added and fixes to make them clear

Credit to ChatGPT

* Make comm parameters intent in

* A few more places

* Remove ' from comments, because that's apparently a thing...

* More places with '

* Turn off Werror

Apparently the version of gfortran on github CI is so absurdly broken
that it gives a warning about C++ style comments in a Fortran file...

* I think my build conditionals are wrong

* Maybe we need to reduce the number of threads, CI is slow

* Simplify thread setting

* typo

+++ Source/CPlusPlus/Logging.cc
@@ -13,7 +13,7 @@ void NTPoly::ActivateLogger(bool start_document) {
 
 ////////////////////////////////////////////////////////////////////////////////
 void NTPoly::ActivateLogger(const string file_name, bool start_document) {
-  int string_length = file_name.length();
+  int string_length = static_cast<int>(file_name.length());
   string temp = file_name;
   ActivateLoggerFile_wrp(&start_document, &temp.c_str()[0], &string_length);
 }
